requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==1.24.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing-extensions==4.12.2
gunicorn